Prefer included modules and use a base module
